+++ extensions/badges/src/types.ts
@@ -1,0 +1,31 @@
+type DuplicateAlias = {
+  title: string;
+  hex?: string;
+  guidelines?: string;
+  loc?: Localization;
+};
+
+export type Localization = {
+  [languageCode: string]: string;
+};
+
+export type Aliases = {
+  aka?: string[];
+  dup?: DuplicateAlias[];
+  loc?: Localization;
+};
+
+export type IconData = {
+  title: string;
+  hex: string;
+  source: string;
+  slug: string;
+  guidelines?: string;
+  license?: { type: string; url?: string };
+  aliases?: Aliases;
+};
+
+export type LaunchFromSimpleIconsContext = {
+  launchFromExtensionName: string;
+  icon: IconData;
+};