+++ extensions/google-translate/src/translate.tsx
@@ -5,7 +5,7 @@ import { useDebouncedValue, useSelectedLanguagesSet, useTextState } from "./hook
 import { getLanguageFlag, supportedLanguagesByCode } from "./languages";
 import { LanguageManagerListDropdown } from "./LanguagesManager";
 import { doubleWayTranslate, playTTS } from "./simple-translate";
-import { ConfigurableCopyPasteActions } from "./actions";
+import { ConfigurableCopyPasteActions, OpenOnGoogleTranslateWebsiteAction, ToggleFullTextAction } from "./actions";
 
 export default function Translate(): ReactElement {
   const [selectedLanguageSet] = useSelectedLanguagesSet();
@@ -52,30 +52,14 @@ export default function Translate(): ReactElement {
                 <ActionPanel>
                   <ActionPanel.Section>
                     <ConfigurableCopyPasteActions defaultActionsPrefix="Translation" value={r.translatedText} />
-                    <Action
-                      title="Toggle Full Text"
-                      icon={Icon.Text}
-                      onAction={() => setIsShowingDetail(!isShowingDetail)}
-                    />
+                    <ToggleFullTextAction onAction={() => setIsShowingDetail(!isShowingDetail)} />
                     <Action
                       title="Play Text-To-Speech"
                       icon={Icon.Play}
                       shortcut={{ modifiers: ["cmd"], key: "t" }}
                       onAction={() => playTTS(r.translatedText, r.langTo)}
                     />
-                    <Action.OpenInBrowser
-                      title="Open in Google Translate"
-                      shortcut={{ modifiers: ["opt"], key: "enter" }}
-                      url={
-                        "https://translate.google.com/?sl=" +
-                        r.langFrom +
-                        "&tl=" +
-                        r.langTo +
-                        "&text=" +
-                        encodeURIComponent(debouncedValue) +
-                        "&op=translate"
-                      }
-                    />
+                    <OpenOnGoogleTranslateWebsiteAction translationText={debouncedValue} translation={r} />
                   </ActionPanel.Section>
                 </ActionPanel>
               }
@@ -89,24 +73,8 @@ export default function Translate(): ReactElement {
                   <ActionPanel>
                     <ActionPanel.Section>
                       <ConfigurableCopyPasteActions value={r.pronunciationText} />
-                      <Action
-                        title="Toggle Full Text"
-                        icon={Icon.Text}
-                        onAction={() => setIsShowingDetail(!isShowingDetail)}
-                      />
-                      <Action.OpenInBrowser
-                        title="Open in Google Translate"
-                        shortcut={{ modifiers: ["opt"], key: "enter" }}
-                        url={
-                          "https://translate.google.com/?sl=" +
-                          r.langFrom +
-                          "&tl=" +
-                          r.langTo +
-                          "&text=" +
-                          encodeURIComponent(debouncedValue) +
-                          "&op=translate"
-                        }
-                      />
+                      <ToggleFullTextAction onAction={() => setIsShowingDetail(!isShowingDetail)} />
+                      <OpenOnGoogleTranslateWebsiteAction translationText={debouncedValue} translation={r} />
                     </ActionPanel.Section>
                   </ActionPanel>
                 }