+++ extensions/shell/CHANGELOG.md
@@ -4,6 +4,10 @@
 This version adds the option to run commands in `kitty` terminal. 
 Checkout the preferences and select your favorite console application in the dropdown to get going.
 
+## [Fix] 2024-04-28
+Allow to run alias commands while running the shell command in raycast.
+
+
 ## [Fix] 2024-04-26
 Fix an issue where iTerm opens two windows instead of one when iTerm was not already running.